# openapi/openapi.yml
openapi: 3.1.0
info:
  title: MLB Forecast API
  description: "\n    # MLB Predictions Backend API\n\n    Enterprise-grade MLB game\
    \ prediction platform delivering real-time analytics and machine learning insights.\n\
    \n    ## \U0001F680 Core Features\n\n    ### \u26BE **Data Ingestion Pipeline**\n\
    \    - Real-time game results from official MLB APIs\n    - Historical statistics\
    \ and performance metrics\n    - Live game schedules and league standings\n  \
    \  - Automated data validation and cleansing\n\n    ### \U0001F916 **Advanced\
    \ Machine Learning**\n    - Probabilistic game outcome predictions\n    - Total\
    \ runs and score forecasting\n    - Player performance analytics\n    - Continuous\
    \ model training and optimization\n\n    ### \u26A1 **High-Performance Architecture**\n\
    \    - Multi-layer Redis caching system\n    - Optimized database queries with\
    \ indexing\n    - Asynchronous processing for scalability\n    - Sub-second API\
    \ response times\n\n    ### \U0001F4CA **Analytics & Insights**\n    - Team performance\
    \ trending\n    - Head-to-head historical analysis\n    - Weather impact on game\
    \ outcomes\n    - Injury reports integration\n\n    ## \U0001F527 Technical Stack\n\
    \n    - **Framework**: FastAPI with async/await patterns\n    - **Database**:\
    \ PostgreSQL with optimized schemas\n    - **Cache**: Redis for high-speed data\
...
    \    ```\n    "
  version: 1.0.0
paths:
...
  /api/v1/predictions/{game_id}:
    get:
      tags:
        - Predictions
      summary: Get Game Predictions
      description: "Get predictions for a specific game.\n\nRetrieves machine learning\
        \ predictions for a game including win probabilities\nand predicted scores\
        \ for both teams.\n\nArgs:\n    game_id: The ID of the game to get predictions\
        \ for\n    prediction_type: Optional filter by prediction type\n    use_cases:\
        \ Dependency injected dictionary containing use cases\n\nReturns:\n    JSONResponse:\
        \ Standardized response with prediction data\n\nRaises:\n    DomainExceptions.InvalidDataError:\
        \ If game_id is invalid"
      operationId: get_game_predictions_api_v1_predictions__game_id__get
      parameters:
        - name: game_id
          in: path
          required: true
          schema:
            type: integer
            description: The ID of the game to get predictions for
            title: Game Id
          description: The ID of the game to get predictions for
        - name: prediction_type
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            description: Filter by prediction type
            title: Prediction Type
          description: Filter by prediction type
      responses:
        '200':
          description: Successfully retrieved predictions for game
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictionListResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
                description: Bad Request
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
                description: Not Found
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnprocessableEntity'
                description: Unprocessable Entity
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
                description: Internal Server Error
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceUnavailable'
                description: Service Unavailable
  /api/v1/predictions:
    post:
      tags:
        - Predictions
      summary: Generate Prediction
      description: "Generate a new prediction for a game.\n\nCreates machine learning\
        \ predictions for a game using current team statistics\nand historical performance\
        \ data.\n\nArgs:\n    game_id: The ID of the game to create prediction for\n\
        \    prediction_type: Type of prediction to generate (winner, total_runs,\
        \ etc.)\n    use_cases: Dependency injected dictionary containing use cases\n\
        \nReturns:\n    JSONResponse: Standardized response with created prediction\n\
        \nRaises:\n    DomainExceptions.GameNotFoundError: If the game is not found\n\
        \    DomainExceptions.InvalidDataError: If game_id is invalid\n    DomainExceptions.ExternalServiceError:\
        \ If ML model is unavailable"
      operationId: generate_prediction_api_v1_predictions_post
      parameters:
        - name: game_id
          in: query
          required: true
          schema:
            type: integer
            description: The ID of the game to create prediction for
            title: Game Id
          description: The ID of the game to create prediction for
        - name: prediction_type
          in: query
          required: false
          schema:
            type: string
            description: Type of prediction to generate
            default: winner
            title: Prediction Type
          description: Type of prediction to generate
      responses:
        '201':
          description: Successfully created prediction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictionCreateResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
                description: Bad Request
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
                description: Not Found
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnprocessableEntity'
                description: Unprocessable Entity
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
                description: Internal Server Error
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceUnavailable'
                description: Service Unavailable
...
components:
  schemas:
...
    BadRequest:
      description: Bad Request - Invalid input parameters
      title: Bad Request
      $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not Found - Resource not found
      title: Not Found
      $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Validation Error - Request validation failed
      title: Validation Error
      $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal Server Error
      title: Internal Server Error
      $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: Service Unavailable - External service error
      title: Service Unavailable
      $ref: '#/components/schemas/ErrorResponse'
    ErrorResponse:
      properties:
        status:
          type: string
          title: Status
          description: Response status
          default: error
        code:
          type: integer
          title: Code
          description: HTTP status code
        data:
          type: 'null'
          title: Data
          description: No data in error responses
        errors:
          items:
            type: string
          type: array
          title: Errors
          description: List of error messages
        message:
          anyOf:
            - type: string
            - type: 'null'
          title: Message
          description: Error message
      type: object
      required:
        - code
        - errors
      title: ErrorResponse
      description: Error response model for API documentation.
      examples:
        - code: 400
          errors:
            - League must be 'American' or 'National'
          message: Invalid data provided
          status: error
        - code: 404
          errors:
            - Team with ID 999 not found
          message: Resource not found
          status: error
        - code: 500
          errors:
            - An unexpected error occurred
          message: Internal server error
          status: error
...
    PredictionCreateData:
      properties:
        prediction:
          $ref: '#/components/schemas/PredictionDTO'
          description: Created prediction details
        processing_time_seconds:
          type: number
          title: Processing Time Seconds
          description: Time taken to generate prediction
        created_at:
          type: string
          title: Created At
          description: Creation timestamp
      type: object
      required:
        - prediction
        - processing_time_seconds
        - created_at
      title: PredictionCreateData
      description: Data model for prediction creation response.
    PredictionCreateResponse:
      properties:
        status:
          type: string
          title: Status
          default: success
        code:
          type: integer
          title: Code
          description: HTTP status code
        data:
          $ref: '#/components/schemas/PredictionCreateData'
          description: Prediction creation details
        errors:
          items:
            type: string
          type: array
          title: Errors
          description: List of error messages
        message:
          type: string
          title: Message
          description: Success message
      type: object
      required:
        - code
        - data
        - message
      title: PredictionCreateResponse
      description: Response model for prediction creation endpoints.
      example:
        status: success
        code: 201
        data:
          prediction:
            id: 1
            game_id: 12345
            model_version: v1.2.0
            home_win_probability: 0.62
            away_win_probability: 0.38
            predicted_home_score: 5.2
            predicted_away_score: 3.8
            confidence_score: 0.75
            created_at: '2025-07-25T18:30:15.123456Z'
          processing_time_seconds: 1.23
          created_at: '2025-07-25T18:30:15.123456Z'
        errors: [ ]
        message: Prediction created successfully for game 12345
    PredictionDTO:
      properties:
        id:
          anyOf:
            - type: integer
            - type: 'null'
          title: Id
          description: Internal prediction ID
        game_id:
          type: integer
          title: Game Id
          description: Associated game ID
        ml_model_version:
          type: string
          title: Ml Model Version
          description: ML model version used
        home_win_probability:
          type: number
          title: Home Win Probability
          description: Probability of home team winning
        away_win_probability:
          type: number
          title: Away Win Probability
          description: Probability of away team winning
        predicted_home_score:
          anyOf:
            - type: number
            - type: 'null'
          title: Predicted Home Score
          description: Predicted home team score
        predicted_away_score:
          anyOf:
            - type: number
            - type: 'null'
          title: Predicted Away Score
          description: Predicted away team score
        confidence_score:
          type: number
          title: Confidence Score
          description: Prediction confidence level
        created_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Created At
          description: Prediction timestamp
      type: object
      required:
        - game_id
        - ml_model_version
        - home_win_probability
        - away_win_probability
        - confidence_score
      title: PredictionDTO
      description: Data Transfer Object for Prediction entity.
    PredictionListResponse:
      properties:
        status:
          type: string
          title: Status
          default: success
        code:
          type: integer
          title: Code
          description: HTTP status code
        data:
          items:
            $ref: '#/components/schemas/PredictionDTO'
          type: array
          title: Data
          description: List of predictions
        errors:
          items:
            type: string
          type: array
          title: Errors
          description: List of error messages
        message:
          type: string
          title: Message
          description: Success message
      type: object
      required:
        - code
        - data
        - message
      title: PredictionListResponse
      description: Response model for prediction list endpoints.
      example:
        status: success
        code: 200
        data:
          - id: 1
            game_id: 12345
            model_version: v1.2.0
            home_win_probability: 0.62
            away_win_probability: 0.38
            predicted_home_score: 5.2
            predicted_away_score: 3.8
            confidence_score: 0.75
            created_at: '2025-07-25T18:30:15.123456Z'
        errors: [ ]
        message: Retrieved 1 predictions for game 12345
...